mame.lst: Added "racedrivpab"
Added "racedrivpab" which is a British Race Drivin' Panorama frankenstein ROM i made by mixing the 2 British Race Drivin' compact ROM's responsible for making the car right hand drive and the Street Drivin'  "mainpcb:user1" ROM's. It excludes the ROM's for the left and right monitor cabinets. 

The result is a British version of Race Drivin' Panorama that runs on a single screen with improved frame rate because MAME is only emulating 1 cabinet instead of 3.